ONNX_learning/onnx_requirements.txt
onnxsim
torchvision
onnxruntime
safetensors
accelerate
onnxoptimizer